refact codes
